requirements.txt
numpy
streamlit
google-generativeai
hydralit-components
streamlit_option_menu
streamlit_lottie
